Update sample.html
I used new Date() to instantiate July 4, 1776. Then I used .getDay() to get the day of week for said date. Using the already existing HTML document, I switched out the Script for "Hello World" and placed the new script so that it returns the day in which July 4, 1776 took place.