Create homework.html
I ask for user birthdate in Short Date Format. The I subtract the birthDay from the currentDate and this gives me the difference in Miliseconds. To convert to days, I use an equation to divide it by the product of multiplying 1000 (miliseconds in a second) times 60 (seconds in a minute) times 60 (minutes in an hour) times 24 (hours in a day). I also added a .floor function to round it down to the nearest integer.