Update to homework1.html & homework2.html
On homework1.html, I converted it to 24hr format.

For homewrok2.html, I created a for loop and placed if statement inside of for loop and adjusted arrays to simplify the code.

Also spaced out commas in array, used let for variables that I'm re-assigning. I'll look into adding an auto formatter on Atom.